feat: Moves file token group into public
