+++ src/file-token-group/thumbnail.tsx
