+++ src/file-upload/internal.tsx
@@ -9,13 +9,13 @@ import { warnOnce } from '@cloudscape-design/component-toolkit/internal';
 import InternalBox from '../box/internal';
 import { ButtonProps } from '../button/interfaces';
 import { useFormFieldContext } from '../contexts/form-field';
-import InternalFileInput from '../file-input/internal';
-import { getBaseProps } from '../internal/base-component';
 import InternalFileDropzone from '../file-dropzone/internal';
 import { useFilesDragging } from '../file-dropzone/use-files-dragging';
+import InternalFileInput from '../file-input/internal';
+import InternalFileToken from '../file-token-group/file-token';
+import InternalFileTokenGroup from '../file-token-group/internal';
 import { ConstraintText, FormFieldError, FormFieldWarning } from '../form-field/internal';
-import InternalFileTokenGroup from '../internal/components/file-token-group';
-import InternalFileToken from '../internal/components/file-token-group/file-token';
+import { getBaseProps } from '../internal/base-component';
 import { fireNonCancelableEvent } from '../internal/events';
 import checkControlled from '../internal/hooks/check-controlled';
 import { InternalBaseComponentProps } from '../internal/hooks/use-base-component';
